+++ packages/cohere/src/cohere-chat-language-model.ts
@@ -21,12 +21,6 @@ import { cohereFailedResponseHandler } from '../src/cohere-error';
 import { convertToCohereChatPrompt } from '../src/convert-to-cohere-chat-prompt';
 import { mapCohereFinishReason } from '../src/map-cohere-finish-reason';
 import { prepareTools } from './cohere-prepare-tools';
-import {
-  CohereChatPrompt,
-  CohereAssistantMessage,
-  CohereUserMessage,
-  CohereSystemMessage,
-} from './cohere-chat-prompt';
 
 type CohereChatConfig = {
   provider: string;
@@ -37,7 +31,7 @@ type CohereChatConfig = {
 
 export class CohereChatLanguageModel implements LanguageModelV1 {
   readonly specificationVersion = 'v1';
-  readonly defaultObjectGenerationMode = undefined;
+  readonly defaultObjectGenerationMode = 'json';
 
   readonly modelId: CohereChatModelId;
   readonly settings: CohereChatSettings;
@@ -79,9 +73,6 @@ export class CohereChatLanguageModel implements LanguageModelV1 {
       // model id:
       model: this.modelId,
 
-      // model specific settings:
-      // none
-
       // standardized settings:
       frequency_penalty: frequencyPenalty,
       presence_penalty: presencePenalty,
@@ -95,7 +86,7 @@ export class CohereChatLanguageModel implements LanguageModelV1 {
       // response format:
       response_format:
         responseFormat?.type === 'json'
-          ? { type: 'json_object', schema: responseFormat.schema }
+          ? { type: 'json_object', json_schema: responseFormat.schema }
           : undefined,
 
       // messages:
@@ -104,27 +95,49 @@ export class CohereChatLanguageModel implements LanguageModelV1 {
 
     switch (type) {
       case 'regular': {
-        const { tools, tool_choice, toolWarnings } = prepareTools(mode);
-        // TODO(shaper): Cohere API doesn't appear to support any form of
-        // explicit tool choice currently. In the future we may want to pass
-        // along the `tool_choice` value in some manner.
+        const { tools, toolChoice, toolWarnings } = prepareTools(mode);
+
         return {
-          ...baseArgs,
-          tools,
+          args: {
+            ...baseArgs,
+            tools,
+            tool_choice: toolChoice,
+          },
           warnings: toolWarnings,
         };
       }
 
       case 'object-json': {
-        throw new UnsupportedFunctionalityError({
-          functionality: 'object-json mode',
-        });
+        return {
+          args: {
+            ...baseArgs,
+            response_format:
+              mode.schema == null
+                ? { type: 'json_object' }
+                : { type: 'json_object', json_schema: mode.schema },
+          },
+          warnings: [],
+        };
       }
 
       case 'object-tool': {
-        throw new UnsupportedFunctionalityError({
-          functionality: 'object-tool mode',
-        });
+        return {
+          args: {
+            ...baseArgs,
+            tools: [
+              {
+                type: 'function',
+                function: {
+                  name: mode.tool.name,
+                  description: mode.tool.description ?? '',
+                  parameters: mode.tool.parameters,
+                },
+              },
+            ],
+            tool_choice: 'REQUIRED',
+          },
+          warnings: [],
+        };
       }
 
       default: {
@@ -136,60 +149,10 @@ export class CohereChatLanguageModel implements LanguageModelV1 {
     }
   }
 
-  concatenateMessageText(messages: CohereChatPrompt): string {
-    return messages
-      .filter(
-        (
-          message,
-        ): message is
-          | CohereSystemMessage
-          | CohereUserMessage
-          | CohereAssistantMessage => 'content' in message,
-      )
-      .map(message => message.content)
-      .join('');
-  }
-
-  /*
-  Remove `additionalProperties` and `$schema` from the `parameters` object of each tool.
-  Though these are part of JSON schema, Cohere chokes if we include them in the request.
-  */
-  // TODO(shaper): Look at defining a type to simplify the params here and a couple of other places.
-  removeJsonSchemaExtras(
-    tools: Array<{
-      type: 'function';
-      function: {
-        name: string | undefined;
-        description: string | undefined;
-        parameters: unknown;
-      };
-    }>,
-  ) {
-    return tools.map(tool => {
-      if (
-        tool.type === 'function' &&
-        tool.function.parameters &&
-        typeof tool.function.parameters === 'object'
-      ) {
-        const { additionalProperties, $schema, ...restParameters } = tool
-          .function.parameters as Record<string, unknown>;
-        return {
-          ...tool,
-          function: {
-            ...tool.function,
-            parameters: restParameters,
-          },
-        };
-      }
-      return tool;
-    });
-  }
-
   async doGenerate(
     options: Parameters<LanguageModelV1['doGenerate']>[0],
   ): Promise<Awaited<ReturnType<LanguageModelV1['doGenerate']>>> {
-    const { warnings, ...args } = this.getArgs(options);
-    args.tools = args.tools && this.removeJsonSchemaExtras(args.tools);
+    const { args, warnings } = this.getArgs(options);
 
     const {
       responseHeaders,
@@ -249,14 +212,12 @@ export class CohereChatLanguageModel implements LanguageModelV1 {
   async doStream(
     options: Parameters<LanguageModelV1['doStream']>[0],
   ): Promise<Awaited<ReturnType<LanguageModelV1['doStream']>>> {
-    const { warnings, ...args } = this.getArgs(options);
-    args.tools = args.tools && this.removeJsonSchemaExtras(args.tools);
-    const body = { ...args, stream: true };
+    const { args, warnings } = this.getArgs(options);
 
     const { responseHeaders, value: response } = await postJsonToApi({
       url: `${this.config.baseURL}/chat`,
       headers: combineHeaders(this.config.headers(), options.headers),
-      body,
+      body: { ...args, stream: true },
       failedResponseHandler: cohereFailedResponseHandler,
       successfulResponseHandler: createEventSourceResponseHandler(
         cohereChatChunkSchema,
@@ -416,7 +377,7 @@ export class CohereChatLanguageModel implements LanguageModelV1 {
       },
       rawResponse: { headers: responseHeaders },
       warnings,
-      request: { body: JSON.stringify(body) },
+      request: { body: JSON.stringify({ ...args, stream: true }) },
     };
   }
 }