+++ packages/provider-utils/src/response-handler.ts
@@ -1,11 +1,11 @@
 import { APICallError, EmptyResponseBodyError } from '@ai-sdk/provider';
-import {
-  EventSourceParserStream,
-  EventSourceMessage,
-} from 'eventsource-parser/stream';
 import { ZodSchema } from 'zod';
+import {
+  createEventSourceParserStream,
+  EventSourceChunk,
+} from './event-source-parser-stream';
 import { extractResponseHeaders } from './extract-response-headers';
-import { ParseResult, parseJSON, safeParseJSON } from './parse-json';
+import { parseJSON, ParseResult, safeParseJSON } from './parse-json';
 
 export type ResponseHandler<RETURN_TYPE> = (options: {
   url: string;
@@ -98,9 +98,9 @@ export const createEventSourceResponseHandler =
       responseHeaders,
       value: response.body
         .pipeThrough(new TextDecoderStream())
-        .pipeThrough(new EventSourceParserStream())
+        .pipeThrough(createEventSourceParserStream())
         .pipeThrough(
-          new TransformStream<EventSourceMessage, ParseResult<T>>({
+          new TransformStream<EventSourceChunk, ParseResult<T>>({
             transform({ data }, controller) {
               // ignore the 'DONE' event that e.g. OpenAI sends:
               if (data === '[DONE]') {