+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,41 @@
 
 ### Patch Changes
 
+- ai@4.2.5
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [3d6d96d]
+  - ai@4.2.4
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0b3bf29]
+  - ai@4.2.3
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [f10f0fa]
+  - ai@4.2.2
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [b796152]
+- Updated dependencies [06361d6]
+  - ai@4.2.1
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [5bc638d]
   - ai@4.2.0