+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/deepseek
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [f10f0fa]
+  - @ai-sdk/provider-utils@2.2.1
+  - @ai-sdk/openai-compatible@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes