+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/openai
 
+## 1.3.2
+
+### Patch Changes
+
+- db15028: feat (provider/openai): expose type for validating OpenAI responses provider options
+
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [f10f0fa]
+  - @ai-sdk/provider-utils@2.2.1
+
 ## 1.3.0
 
 ### Minor Changes